[ refactor ] 서비스 레이어 추상화 ( #38 )
